added the new UI to the common area
